Update README.md (#27)
* Update README.md

* Update code-quality.yml

* Update code-quality.yml

* fix

* Update README.md

* Update README.md

* type casting

+++ app/Services/Models/ExpenseService.php
@@ -57,7 +57,7 @@ class ExpenseService
         ExpenseData $expenseData,
         DateData $dateData
     ): ?float {
-        return Expense::query()
+        return (float) Expense::query()
             ->where('user_id', $expenseData->userId)
             ->whereMonth('spent_at', $dateData->month)
             ->whereYear('spent_at', $dateData->year)
